refactor: reorganized util files under one directory
