+++ src/utils/mod.rs
@@ -1,0 +1,2 @@
+pub mod custom_highlighter;
+pub mod util_functions;
